move service controller config to k8s.io/cloud-provider/controllers/service/config
