+++ cmd/kube-controller-manager/app/options/BUILD
@@ -129,8 +129,8 @@ go_test(
         "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/server/options:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config:go_default_library",
         "//staging/src/k8s.io/cloud-provider/options:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
         "//staging/src/k8s.io/component-base/config:go_default_library",
         "//staging/src/k8s.io/component-base/logs:go_default_library",
         "//staging/src/k8s.io/component-base/metrics:go_default_library",
