+++ pkg/controller/apis/config/BUILD
@@ -36,7 +36,7 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config:go_default_library",
         "//staging/src/k8s.io/controller-manager/config:go_default_library",
     ],
 )
