+++ pkg/controller/apis/config/v1alpha1/doc.go
@@ -18,7 +18,7 @@ limitations under the License.
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/apis/config
 // +k8s:conversion-gen=k8s.io/component-base/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/cloud-provider/config/v1alpha1
-// +k8s:conversion-gen=k8s.io/cloud-provider/service/config/v1alpha1
+// +k8s:conversion-gen=k8s.io/cloud-provider/controllers/service/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/controller-manager/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/certificates/signer/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/daemon/config/v1alpha1
