+++ staging/src/k8s.io/cloud-provider/config/v1alpha1/BUILD
@@ -21,7 +21,7 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config/v1alpha1:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config/v1alpha1:go_default_library",
         "//staging/src/k8s.io/controller-manager/config/v1alpha1:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
     ],
