+++ staging/src/k8s.io/cloud-provider/config/v1alpha1/doc.go
@@ -23,7 +23,7 @@ limitations under the License.
 // +k8s:conversion-gen=k8s.io/component-base/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/cloud-provider/config
 // +k8s:conversion-gen=k8s.io/cloud-provider/config/v1alpha1
-// +k8s:conversion-gen=k8s.io/cloud-provider/service/config/v1alpha1
+// +k8s:conversion-gen=k8s.io/cloud-provider/controllers/service/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/controller-manager/config/v1alpha1
 // +k8s:openapi-gen=true
 // +k8s:defaulter-gen=TypeMeta
