+++ staging/src/k8s.io/cloud-provider/controllers/service/config/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // +k8s:deepcopy-gen=package
 
-package config // import "k8s.io/cloud-provider/service/config"
+package config // import "k8s.io/cloud-provider/controllers/service/config"
