+++ staging/src/k8s.io/cloud-provider/controllers/service/config/v1alpha1/BUILD
@@ -11,13 +11,13 @@ go_library(
         "zz_generated.conversion.go",
         "zz_generated.deepcopy.go",
     ],
-    importmap = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/service/config/v1alpha1",
-    importpath = "k8s.io/cloud-provider/service/config/v1alpha1",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/controllers/service/config/v1alpha1",
+    importpath = "k8s.io/cloud-provider/controllers/service/config/v1alpha1",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config:go_default_library",
     ],
 )
