+++ staging/src/k8s.io/cloud-provider/controllers/service/config/v1alpha1/doc.go
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 // +k8s:deepcopy-gen=package
-// +k8s:conversion-gen=k8s.io/cloud-provider/service/config
-// +k8s:conversion-gen=k8s.io/cloud-provider/service/config/v1alpha1
+// +k8s:conversion-gen=k8s.io/cloud-provider/controllers/service/config
+// +k8s:conversion-gen=k8s.io/cloud-provider/controllers/service/config/v1alpha1
 
-package v1alpha1 // import "k8s.io/cloud-provider/service/config/v1alpha1"
+package v1alpha1 // import "k8s.io/cloud-provider/controllers/service/config/v1alpha1"
