+++ staging/src/k8s.io/cloud-provider/options/BUILD
@@ -29,7 +29,7 @@ go_library(
         "//staging/src/k8s.io/cloud-provider/config:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config/install:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config/v1alpha1:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
         "//staging/src/k8s.io/controller-manager/options:go_default_library",
         "//staging/src/k8s.io/controller-manager/pkg/clientbuilder:go_default_library",
@@ -61,7 +61,7 @@ go_test(
         "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/server/options:go_default_library",
         "//staging/src/k8s.io/cloud-provider/config:go_default_library",
-        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/service/config:go_default_library",
         "//staging/src/k8s.io/component-base/config:go_default_library",
         "//staging/src/k8s.io/controller-manager/config:go_default_library",
         "//staging/src/k8s.io/controller-manager/options:go_default_library",
